[xy] Add base http client.
